+++ src/bacnet/basic/object/piv.h
@@ -22,8 +22,9 @@ extern "C" {
 
 typedef struct positiveinteger_value_descr {
     bool Out_Of_Service : 1;
-    uint32_t Present_Value;
+    BACNET_UNSIGNED_INTEGER Present_Value;
     BACNET_ENGINEERING_UNITS Units;
+    const char *Object_Name;
 } POSITIVEINTEGER_VALUE_DESCR;
 
 BACNET_STACK_EXPORT
@@ -46,6 +47,11 @@ unsigned PositiveInteger_Value_Instance_To_Index(uint32_t object_instance);
 BACNET_STACK_EXPORT
 bool PositiveInteger_Value_Object_Name(
     uint32_t object_instance, BACNET_CHARACTER_STRING *object_name);
+BACNET_STACK_EXPORT
+bool PositiveInteger_Value_Name_Set(
+    uint32_t object_instance, const char *new_name);
+BACNET_STACK_EXPORT
+const char *PositiveInteger_Value_Name_ASCII(uint32_t object_instance);
 
 BACNET_STACK_EXPORT
 int PositiveInteger_Value_Read_Property(BACNET_READ_PROPERTY_DATA *rpdata);
@@ -55,9 +61,10 @@ bool PositiveInteger_Value_Write_Property(BACNET_WRITE_PROPERTY_DATA *wp_data);
 
 BACNET_STACK_EXPORT
 bool PositiveInteger_Value_Present_Value_Set(
-    uint32_t object_instance, uint32_t value, uint8_t priority);
+    uint32_t object_instance, BACNET_UNSIGNED_INTEGER value, uint8_t priority);
 BACNET_STACK_EXPORT
-uint32_t PositiveInteger_Value_Present_Value(uint32_t object_instance);
+BACNET_UNSIGNED_INTEGER
+PositiveInteger_Value_Present_Value(uint32_t object_instance);
 
 BACNET_STACK_EXPORT
 bool PositiveInteger_Value_Change_Of_Value(uint32_t instance);
@@ -78,8 +85,11 @@ bool PositiveInteger_Value_Out_Of_Service(uint32_t instance);
 BACNET_STACK_EXPORT
 void PositiveInteger_Value_Out_Of_Service_Set(uint32_t instance, bool oos_flag);
 
-/* note: header of Intrinsic_Reporting function is required
-   even when INTRINSIC_REPORTING is not defined */
+BACNET_STACK_EXPORT
+uint32_t PositiveInteger_Value_Create(uint32_t object_instance);
+BACNET_STACK_EXPORT
+bool PositiveInteger_Value_Delete(uint32_t object_instance);
+
 BACNET_STACK_EXPORT
 void PositiveInteger_Value_Intrinsic_Reporting(uint32_t object_instance);